build 2.0

+++ src/app/prod-show/prod-show.component.ts
@@ -17,8 +17,9 @@ import { Ng2PicaService } from 'ng2-pica';
   styleUrls: ['./prod-show.component.sass']
 })
 export class ProdShowComponent implements OnInit {
+  @ViewChild('carousel') carouselElement; 
+  actions = new EventEmitter<string>();
 	id: number;
-  images:any = [];
   modalActions = new EventEmitter<string|MaterializeAction>();
   AWS = require('aws-sdk');
   img_upload:boolean = false;
@@ -26,6 +27,9 @@ export class ProdShowComponent implements OnInit {
   prod = new Prod;
   comment = new Comment;
   err:string;
+  images:any[] = [[]];
+  sucess:boolean = false;
+
 
 
   constructor(
@@ -40,40 +44,52 @@ export class ProdShowComponent implements OnInit {
     this.route.params.subscribe(
       params => {this.id = params['id']}
     )
+    let self = this;
     let timer = Observable.timer(0, 5000);
     timer.subscribe(() =>{
       this.prodService.getProd(this.id).subscribe(res=>{
-        this.prod = res.json().prod;
         this.comments = res.json().comments;
-        this.images = this.prod.image.split(',');
       });
     });
+    this.prodService.getProd(this.id).subscribe(res=>{
+      this.prod = res.json().prod;
+      let a = 0;
+      let b = 0;
+      this.prod.image.split(',').forEach(function(el){
+        if (a < 4){self.images[b][a] = el;a+=1}else{b+=1;a=1;self.images[b] = [el]}
+      })
+      this.carouselElement.nativeElement.classList.toggle('initialized');
+      this.actions.emit('carousel');
+    })
   }
-
-  // @Input() prod: Prod;
 
   ngOnInit() {
   }
 
   deleteProd(id:number){
-    let self = this;
-    this.AWS.config.update({region: 'ap-southeast-1', credentials: {"accessKeyId": this.prodService.env.id, "secretAccessKey": this.prodService.env.key}});
-    let a = this.prod.image.split(',');
-    a.forEach(function(element){
-      let s3 = new self.AWS.S3().deleteObject({Bucket: self.prodService.env.bucket, Key: element},function(err, data) {
-        if (err) {console.log(err, err.stack);} // an error occurred
-        else     {console.log(data);}         // successful response
-      }); 
-    });
-  	this.prodService.deleteProd(id)
-  	  .subscribe(res=>{console.log(res); this.router.navigate(['/prods'])});
-    this.bucket.getBucket().subscribe(res=>{
-      let a = [];
-      res.json().forEach(function(el){
-        a.push(el.number);
+    if(window.confirm('Are sure you want to delete this item ?')){
+      let self = this;
+      this.AWS.config.update({region: 'ap-southeast-1', credentials: {"accessKeyId": this.prodService.env.id, "secretAccessKey": this.prodService.env.key}});
+      let a = this.prod.image.split(',');
+      a.forEach(function(element){
+        let s3 = new self.AWS.S3().deleteObject({Bucket: self.prodService.env.bucket, Key: element},function(err, data) {
+          if (err) {} // an error occurred
+          else     {}         // successful response
+        }); 
       });
-      this.bucket.amount = a.reduce((a, b) => a + b, 0);
-    });
+    	this.prodService.deleteProd(id)
+    	  .subscribe(res=>{
+          this.prodService.prods = res.json();
+          this.router.navigate(['/prods']);
+        });
+      this.bucket.getBucket().subscribe(res=>{
+        let a = [];
+        res.json().forEach(function(el){
+          a.push(el.number);
+        });
+        this.bucket.amount = a.reduce((a, b) => a + b, 0);
+      });
+    }
   }
 
   openModal() {
@@ -87,8 +103,16 @@ export class ProdShowComponent implements OnInit {
   update(prod: Prod){
     this.prodService.updateProd(prod)
       .subscribe(data => {
+        let self = this;
         this.prod = data;
-        this.images = this.prod.image.split(',');
+        let a = 0;
+        let b = 0;
+        this.images = [[]];
+        this.prod.image.split(',').forEach(function(el){
+          if (a < 4){self.images[b][a] = el;a+=1}else{b+=1;a=1;self.images[b] = [el]}
+        })
+        self.carouselElement.nativeElement.classList.toggle('initialized');
+        self.actions.emit('carousel');
       },
         error => {
           console.log(error)
@@ -107,6 +131,7 @@ export class ProdShowComponent implements OnInit {
 
   fileEvent(data:any){
     let self = this;
+    this.sucess = false;
     this.img_upload = true;
     let params = {Bucket: this.prodService.env.bucket, Key: this.prod.id+'/'+this.makeid(), Body: data};
     let s3 = new this.AWS.S3.ManagedUpload({params: params});
@@ -114,10 +139,20 @@ export class ProdShowComponent implements OnInit {
       $('#pus').css('width','0%');
       $('#pus').css('width',evt.loaded*100/evt.total+'%');
     }).send(function(error, s3res) { 
-      if (error) {self.err = error.message;}else{};
+      if (error) {self.err = error.message;}else{self.sucess = true;};
       let img:string;
       if(self.prod.image){img = ','+s3res.Key}else{img = s3res.Key};
-      self.prodService.addFile(img, self.prod.id).subscribe(res => {self.prod = res.json();self.images = self.prod.image.split(',');});
+      self.prodService.addFile(img, self.prod.id).subscribe(res => {
+        self.prod = res.json();
+        let a = 0;
+        let b = 0;
+        self.images = [[]];
+        self.prod.image.split(',').forEach(function(el){
+          if (a < 4){self.images[b][a] = el;a+=1}else{b+=1;a=1;self.images[b] = [el]}
+        })
+        self.carouselElement.nativeElement.classList.toggle('initialized');
+        self.actions.emit('carousel');
+      });
       
     });
   }
@@ -133,14 +168,26 @@ export class ProdShowComponent implements OnInit {
   }
 
   delImg(img:string){
+    console.log(img); 
     let self = this;
     this.AWS.config.update({region: 'ap-southeast-1', credentials: {"accessKeyId": this.prodService.env.id, "secretAccessKey": this.prodService.env.key}});
     let s3 = new this.AWS.S3().deleteObject({Bucket: this.prodService.env.bucket, Key: img},function(err, data) {
-      if (err) {console.log(err, err.stack);} // an error occurred
+      if (err) {console.log('im here@@@@@@@@@@@@@@@@',err)} // an error occurred
       else     {
-        console.log(data); 
-        self.prodService.delFile(img, self.prod.id).subscribe(res => {self.prod = res.json();self.images = self.prod.image.split(',');});
-        }         // successful response
+        // console.log()
+        self.prodService.delFile(img, self.prod.id).subscribe(res => {
+          self.prod = res.json();
+          
+          let a = 0;
+          let b = 0;
+          self.images = [[]];
+          self.prod.image.split(',').forEach(function(el){
+            if (a < 4){self.images[b][a] = el;a+=1}else{b+=1;a=1;self.images[b] = [el]}
+          })
+          self.carouselElement.nativeElement.classList.toggle('initialized');
+          self.actions.emit('carousel');
+        });
+        }         
     }); 
   }
 
@@ -158,10 +205,6 @@ export class ProdShowComponent implements OnInit {
     this.prodService.deleteComment(comment.id, this.prod.id).subscribe(res=>this.comments=res.json());
   }
 
-  goToPicture(prod:Prod){
-    this.router.navigate(['/picture',this.prod.id],{queryParams:{image:prod.image}});
-  }
-
   addToBucket(id:number){
     this.bucket.addToBucket(id).subscribe(res=>{
       this.bucket.buckets = res.json();
@@ -173,10 +216,9 @@ export class ProdShowComponent implements OnInit {
     });
   }
 
-  addclass(event){
-    event.target.classList.toggle('spinable');
-    window.setTimeout(() => {
-      event.target.classList.remove('spinable');
-    }, 1000);
+  flooring(number){
+    return Math.floor(12/number);
   }
+
+
 }
